Fixed text rendering.  The problem was that the third parameter--which tells OpenGL the components of data you are passing in--was '1' when it should have been 'GL_ALPHA'.
While '1' is a legal value to place here it implicitly says the component you are
passing in is the luminance component.  Since we were telling the driver to
store only GL_ALPHA components (the seventh parameter) it will just throw
away any luminance data we pass into it.  Only if we pass in GL_ALPHA data (and
indicate such by passing in GL_ALPHA as the third parameter) will
the OpenGL driver actually pay attention to our data.