Radiosity rendering for crystal space. Set the NEWRADIOSITY=yes in cryst.cfg to enable it. Caveat: it is not stored in precalc.info, so always use -recalc. Note: flarge takes an hour with 64x64 lightgrid for me. Reduce your ambient white before using radiosity - it will light up as well. Remove 20 to 40 from it.
The handbooks say that a small patch size is necessary, thus 16 is 'okay'.
But time increases ^2 or ^3 in the number of patches... So beware.

A feature is needed to really use radiosity: ambient_color for a polygon,
so that a polygon is 'self lit', radiosity will then spread the light
to surroundings.

The global ambient color is removed before radiosity calculation, and
put back afterwards.

Also the old pseudo-radiosity algorithm is not useless, it is a global-
reflection algorithm, and having a 'reflection' fraction per polygon
would be nice, as well as the global_reflection.

When you're adding params to polygons add 'diffuse' for the diffuse
reflection used by the radiosity algorithm. 0.7 now, not often changed.

Have fun,
		Wouter