made the following changes to the typed vector system: - moved all macros to their own header file, typedvec.h - internally, an even more flexible macro is used that allows usage of a   user-defined superclass for the vector - modified the vectors for SCF objects. The contained class no longer needs   to be defined at the time the vector is declared. On the bad side, it is now   a requirement that the contained class must be a subclass of iBase. It is   no longer sufficient that the class has IncRef/DecRef methods. However,   this should not be too much of a limitation. - vectors for normal objects and for SCF objects again use the same concept   in their macro hierarchy
added a new system of object vectors:
- This system should finally replace csNamedObjVector. The common
superclass for those vectors is currently called csNamedObjectVector and
is kept in the same source/header files. It extends the macros for typed
vectors to produce type-safe object vectors.
- In addition, csNamedObjVector itself provides all methods that are safe even
if you don't know the type of contained objects. In this case, the objects are
exposed to the user as (iObject*) pointers. This way you can deal with
object vectors that store objects of unknown type, although the range of
possible operations is of course limited.

Finally, I re-implemented csTextureList using typed object vectors. Seems to
work so far.