Did a pedantic change to PI and PI / 2 :-)
Removed over use of variables in "for" loops within fuctions. This is well illustrated
by the amount of "int i"'s there were in "protex2d.cpp" :-)

Did a little bit of loop optimization.

Bug to be verified:

In csvid->opengl, are videos palying upside down for others are they are for me?