Add m4 utilities for reporting a summary of notable configure checks.
CS_NOTABLE() takes note of the result of checks considered significant
enough to appear in an end-of-configure summary report.  In idiomatic
usage, it wraps each check to which it is applied, although it also can
be invoked after-the-fact to record the result of a check.
CS_SUMMARIZE() generates a report indicating success and failure of
checks deemed notable via CS_NOTABLE().