Revert "masterheaders: Regenerate following addition of csutil/vararg.h."
Apple's llvm-gcc's support for C++ is incomplete. It throws an error
upon encountering a virtual vararg method in a virtual base class. The
implemented work-around for this problem is to #ifdef-out the offending
methods and instead have clients invoke the va_list counterpart of the
vararg method via CS::va_call(). This approach, however, is somewhat
maintenance heavy and just plain ugly. A more maintainable and less
invasive work-around is instead to implement the vararg methods as thin
inline non-virtual wrappers which delegate to the va_list counterparts
automatically.  In preparation for this revised solution, revert the
CS::va_call() approach.