jobber/snapshot: Resolve SVN repository version mismatch.
SourceForge has upgraded the CS/CEL repositories to SVN format 4[1].
The local rsync'd mirrors on crystalspace3d.org from which jobber and
snapshot perform checkouts thus also becomes format 4.  Unfortunately,
Subversion installed on crystalspace3d.org is 1.5.x which does not
understand format 4. Work around this shortcoming by employing a locally
built Subversion 1.7.x.  This updated Subversion is available only
locally in the "crystal" user's PATH; it is not installed in /usr/bin.
In order to employ the updated Subversion, stop using a hardcoded path
to the 'svn' binary, so that the one in PATH is found instead.

[1]: http://serverfault.com/questions/277441/difference-between-the-format-and-db-format-files-in-a-subversion-repository